Replaced Option with Result
Added a Result type. Replaced all Option values with Result. Also created a Error type
so in a case of an error the CLI can create an error message on what went wrong.

Retry is now included and part of the Result type. It retries based on Error. There exists
a function to convert a function throwing Exceptions into a function returning
a Result instead.